requirements.txt
Click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy~=1.19.4
redis~=3.5.3
rq==1.7.0
Werkzeug==1.0.1
pandas==1.1.5
matplotlib==3.3.3
seaborn~=0.11.0
lifelines~=0.25.7
scipy~=1.5.4
statsmodels~=0.12.1
psycopg2-binary==2.8.6
pycodestyle
flake8
pyyaml==5.3.1